backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-karaoke-backend"
version = "0.1.0"
description = "FastAPI backend for AI Karaoke Video Generator (Croonify clone MVP)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.7.0",
  "moviepy>=1.0.3",
  "numpy>=1.24.0",
  "srt>=3.5.3",
  "pydub>=0.25.1",
  "librosa>=0.10.2.post1",
  "openai-whisper>=20231117",
  "aeneas>=1.7.3.0"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "httpx>=0.27.0"
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000


